+++ plugins/autocast/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @flatfile/plugin-autocast
 
+## 0.2.0
+
+### Minor Changes
+
+- 966cf91: This release introduces field filters, and enables filtering by either sheetId or sheetSlug.
+
+## 0.1.0
+
+### Minor Changes
+
+- 15dc4fe: Introduction of the autocast plugin!
+
+### Patch Changes
+
+- Updated dependencies [15dc4fe]
+  - @flatfile/plugin-record-hook@1.0.4
+
 ## 0.0.4
 
 ### Patch Changes