+++ plugins/xml-extractor/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @flatfile/plugin-dxp-config
 
+## 0.4.0
+
+### Minor Changes
+
+- 447691a: Add configurable batch options
+
+### Patch Changes
+
+- Updated dependencies [447691a]
+  - @flatfile/util-extractor@0.3.0
+
+## 0.3.3
+
+### Patch Changes
+
+- @flatfile/util-extractor@0.2.4
+
 ## 0.3.2
 
 ### Patch Changes