+++ utils/extractor/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @flatfile/plugin-extractor-utils
 
+## 0.3.0
+
+### Minor Changes
+
+- 447691a: Add configurable batch options
+
+## 0.2.4
+
+### Patch Changes
+
+- Updated dependencies [69fd41c]
+  - @flatfile/util-common@0.1.0
+
 ## 0.2.3
 
 ### Patch Changes